+++ server/keyshare/keyshareserver/server.go
@@ -121,6 +121,16 @@ func (s *Server) Handler() http.Handler {
 		router.Route("/api/v1", func(r chi.Router) {
 			s.routeHandler(r)
 		})
+
+		router.Route("/api/v2", func(r chi.Router) {
+			// Keyshare sessions with provably secure keyshare protocol
+			r.Use(s.userMiddleware)
+			r.Use(s.authorizationMiddleware)
+			r.Post("/prove/getPs", s.handlePs)
+			r.Post("/prove/getCommitments", s.handleCommitmentsV2)
+			r.Post("/prove/getResponse", s.handleResponseV2)
+			r.Post("/prove/getResponseLinkable", s.handleResponseV2Linkable)
+		})
 	})
 
 	// IRMA server for issuing myirma credential during registration
@@ -182,6 +192,56 @@ func (s *Server) loadIdemixKeys(conf *irma.Configuration) error {
 	return errs.ErrorOrNil()
 }
 
+// /prove/getPs
+func (s *Server) handlePs(w http.ResponseWriter, r *http.Request) {
+	// Fetch from context
+	user := r.Context().Value("user").(*User)
+	authorization := r.Context().Value("authorization").(string)
+
+	// Read keys
+	var keys []irma.PublicKeyIdentifier
+	if err := server.ParseBody(r, &keys); err != nil {
+		server.WriteError(w, server.ErrorInvalidRequest, err.Error())
+		return
+	}
+	if len(keys) == 0 {
+		s.conf.Logger.Info("Malformed request: no keys for P specified")
+		server.WriteError(w, server.ErrorInvalidRequest, "no key specified")
+		return
+	}
+
+	ps, err := s.generatePs(user, authorization, keys)
+	if err != nil && err == keysharecore.ErrInvalidJWT {
+		server.WriteError(w, server.ErrorInvalidRequest, err.Error())
+		return
+	}
+	if err != nil {
+		// already logged
+		server.WriteError(w, server.ErrorInternal, err.Error())
+		return
+	}
+
+	server.WriteJson(w, ps)
+}
+
+func (s *Server) generatePs(user *User, authorization string, keys []irma.PublicKeyIdentifier) (*irma.PMap, error) {
+	// Generate Ps
+	ps, err := s.core.GeneratePs(keysharecore.UserSecrets(user.Secrets), authorization, keys)
+	if err != nil {
+		s.conf.Logger.WithField("error", err).Warn("Could not generate Ps for request")
+		return nil, err
+	}
+
+	// Prepare output message format
+	mappedPs := map[irma.PublicKeyIdentifier]*big.Int{}
+	for i, keyID := range keys {
+		mappedPs[keyID] = ps[i]
+	}
+
+	// And send response
+	return &irma.PMap{Ps: mappedPs}, nil
+}
+
 // /prove/getCommitments
 func (s *Server) handleCommitments(w http.ResponseWriter, r *http.Request) {
 	// Fetch from context
@@ -196,7 +256,7 @@ func (s *Server) handleCommitments(w http.ResponseWriter, r *http.Request) {
 	}
 	if len(keys) == 0 {
 		s.conf.Logger.Info("Malformed request: no keys for commitment specified")
-		server.WriteError(w, server.ErrorInvalidRequest, "No key specified")
+		server.WriteError(w, server.ErrorInvalidRequest, "no key specified")
 		return
 	}
 
@@ -238,6 +298,69 @@ func (s *Server) generateCommitments(user *User, authorization string, keys []ir
 
 	// And send response
 	return &irma.ProofPCommitmentMap{Commitments: mappedCommitments}, nil
+}
+
+// /api/v2/prove/getCommitments
+func (s *Server) handleCommitmentsV2(w http.ResponseWriter, r *http.Request) {
+	// Fetch from context
+	user := r.Context().Value("user").(*User)
+	authorization := r.Context().Value("authorization").(string)
+
+	// Read keys
+	var req irma.GetCommitmentsRequest
+	if err := server.ParseBody(r, &req); err != nil {
+		server.WriteError(w, server.ErrorInvalidRequest, err.Error())
+		return
+	}
+	if len(req.Keys) == 0 {
+		s.conf.Logger.Info("Malformed request: no keys for commitment specified")
+		server.WriteError(w, server.ErrorInvalidRequest, "no key specified")
+		return
+	}
+
+	commitments, err := s.generateCommitmentsV2(user, authorization, req)
+	if err != nil && (err == keysharecore.ErrInvalidChallenge || err == keysharecore.ErrInvalidJWT) {
+		server.WriteError(w, server.ErrorInvalidRequest, err.Error())
+		return
+	}
+	if err != nil {
+		// already logged
+		server.WriteError(w, server.ErrorInternal, err.Error())
+		return
+	}
+
+	server.WriteJson(w, commitments)
+}
+
+func (s *Server) generateCommitmentsV2(user *User, authorization string, req irma.GetCommitmentsRequest) (*irma.ProofPCommitmentMapV2, error) {
+	// Generate commitments
+	commitments, commitID, err := s.core.GenerateCommitments(keysharecore.UserSecrets(user.Secrets), authorization, req.Keys)
+	if err != nil {
+		s.conf.Logger.WithField("error", err).Warn("Could not generate commitments for request")
+		return nil, err
+	}
+
+	// Prepare output message format
+	mappedCommitments := map[irma.PublicKeyIdentifier]*big.Int{}
+	for i, keyID := range req.Keys {
+		mappedCommitments[keyID] = commitments[i].Pcommit
+	}
+
+	// Store needed data for later requests.
+	// Of all keys involved in the current session, store the ID of the last one to be used when
+	// the user comes back later to retrieve her response. gabi.ProofP.P will depend on this public
+	// key, which is used only during issuance. Thus, this assumes that during issuance, the user
+	// puts the key ID of the credential(s) being issued at the last index (indeed, the irmaclient
+	// always puts all ProofU's after the ProofD's in the list of proofs it sends to the IRMA
+	// server).
+	s.store.add(user.Username, &session{
+		KeyID:    req.Keys[len(req.Keys)-1],
+		Hw:       req.Hash,
+		CommitID: commitID,
+	})
+
+	// And send response
+	return &irma.ProofPCommitmentMapV2{Commitments: mappedCommitments}, nil
 }
 
 // /prove/getResponse
@@ -300,6 +423,83 @@ func (s *Server) generateResponse(ctx context.Context, user *User, authorization
 	}
 
 	return proofResponse, nil
+}
+
+// /api/v2/prove/getResponse
+func (s *Server) handleResponseV2(w http.ResponseWriter, r *http.Request) {
+	s.keyshareResponse(r.Context(), w, r, false)
+}
+
+func (s *Server) keyshareResponse(ctx context.Context, w http.ResponseWriter, r *http.Request, linkable bool) {
+	// Fetch from context
+	user := r.Context().Value("user").(*User)
+	authorization := r.Context().Value("authorization").(string)
+
+	var req gabi.KeyshareResponseRequest[irma.PublicKeyIdentifier]
+	if err := server.ParseBody(r, &req); err != nil {
+		server.WriteError(w, server.ErrorInvalidRequest, err.Error())
+		return
+	}
+
+	// verify access (avoids leaking whether there is a session ongoing to unauthorized callers)
+	if !r.Context().Value("hasValidAuthorization").(bool) {
+		s.conf.Logger.Warn("Could not generate keyshare response due to invalid authorization")
+		server.WriteError(w, server.ErrorInvalidRequest, "Invalid authorization")
+		return
+	}
+
+	// And do the actual responding
+	proofResponse, err := s.generateResponseV2(ctx, user, authorization, req, linkable)
+	if err != nil &&
+		(err == keysharecore.ErrInvalidChallenge ||
+			err == keysharecore.ErrInvalidJWT ||
+			err == errMissingCommitment) {
+		server.WriteError(w, server.ErrorInvalidRequest, err.Error())
+		return
+	}
+	if err != nil {
+		// already logged
+		server.WriteError(w, server.ErrorInternal, err.Error())
+		return
+	}
+
+	server.WriteString(w, proofResponse)
+}
+
+func (s *Server) generateResponseV2(ctx context.Context, user *User, authorization string, req gabi.KeyshareResponseRequest[irma.PublicKeyIdentifier], linkable bool) (string, error) {
+	// Get data from session
+	sessionData := s.store.get(user.Username)
+	if sessionData == nil {
+		s.conf.Logger.Warn("Request for response without previous call to get commitments")
+		return "", errMissingCommitment
+	}
+
+	// Indicate activity on user account
+	err := s.db.setSeen(ctx, user)
+	if err != nil {
+		s.conf.Logger.WithField("error", err).Error("Could not mark user as seen recently")
+		// Do not send to user
+	}
+
+	// Make log entry
+	err = s.db.addLog(ctx, user, eventTypeIRMASession, nil)
+	if err != nil {
+		s.conf.Logger.WithField("error", err).Error("Could not add log entry for user")
+		return "", err
+	}
+
+	proofResponse, err := s.core.GenerateResponseV2(keysharecore.UserSecrets(user.Secrets), authorization, sessionData.CommitID, sessionData.Hw, req, sessionData.KeyID, linkable)
+	if err != nil {
+		s.conf.Logger.WithField("error", err).Error("Could not generate response for request")
+		return "", err
+	}
+
+	return proofResponse, nil
+}
+
+// /prove/getLinkableResponse
+func (s *Server) handleResponseV2Linkable(w http.ResponseWriter, r *http.Request) {
+	s.keyshareResponse(r.Context(), w, r, true)
 }
 
 // /users/verify_start